# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Jurnapod API",
    "version": "1.0.0",
    "description": "REST API for Jurnapod ERP system"
  },
  "servers": [
    {
      "url": "/api",
      "description": "API Base"
    }
  ],
  "paths": {
    "/outlets/{outletId}/stock": {
      "get": {
        "summary": "Get stock levels",
        "parameters": [
          {
            "name": "outletId",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          },
          {
            "name": "product_id",
            "in": "query",
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Stock levels",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "company_id": { "type": "integer" },
                        "outlet_id": { "type": "integer" },
                        "items": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "product_id": { "type": "integer" },
                              "product_name": { "type": "string" },
                              "quantity": { "type": "integer" },
                              "reserved_quantity": { "type": "integer" },
                              "available_quantity": { "type": "integer" }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    { "bearerAuth": [] }
  ]
}